# backend/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Recipes API
  version: 1.0.0
  description: API para gestión de recetas, ingredientes, menús, usuarios, proveedores, inventario e impuestos.
servers:
  - url: http://localhost:4000
    description: Servidor local de desarrollo
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
    Ingredient:
      type: object
      properties:
        ingredient_id:
          type: integer
        name:
          type: string
        unit:
          type: string
        base_price:
          type: number
        waste_percent:
          type: number
        net_price:
          type: number
        stock:
          type: number
        expiration_date:
          type: string
          format: date
        is_available:
          type: boolean
        comment:
          type: string
    IngredientCreate:
      type: object
      required:
        - name
        - unit
        - base_price
        - waste_percent
        - net_price
        - stock
        - expiration_date
        - is_available
      properties:
        name:
          type: string
        unit:
          type: string
        base_price:
          type: number
        waste_percent:
          type: number
        net_price:
          type: number
        stock:
          type: number
        expiration_date:
          type: string
          format: date
        is_available:
          type: boolean
        comment:
          type: string
    PriceHistoryItem:
      type: object
      properties:
        history_id:
          type: integer
        old_price:
          type: number
        new_price:
          type: number
        change_date:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/UserMini'
    UserMini:
      type: object
      properties:
        user_id:
          type: integer
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
          format: email
    ResponseMessage:
      type: object
      properties:
        message:
          type: string
...
    RecipeIngredient:
      type: object
      properties:
        ingredient_id:
          type: integer
        name:
          type: string
        quantity_per_serving:
          type: number
        unit:
          type: string
        base_price:
          type: number
        waste_percent:
          type: number
        net_price:
          type: number
        section_name:
          type: string
...
    SupplierIngredientItem:
      type: object
      properties:
        ingredient_id:
          type: integer
        name:
          type: string
        price:
          type: number
        delivery_time:
          type: string
          description: Tiempo de entrega (por ejemplo, "3 días")
        is_preferred_supplier:
          type: boolean
    SupplierIngredientCreate:
      type: object
      required:
        - ingredient_id
        - price
        - delivery_time
        - is_preferred_supplier
      properties:
        ingredient_id:
          type: integer
        price:
          type: number
        delivery_time:
          type: string
          description: Tiempo de entrega (por ejemplo, "3 días")
        is_preferred_supplier:
          type: boolean
...
    Allergen:
      type: object
      properties:
        allergen_id:
          type: integer
        name:
          type: string
      required:
        - allergen_id
        - name
...
paths:
...
  /ingredients:
    get:
      tags:
        - Ingredientes
      summary: Listar ingredientes
      description: |
        Obtiene el catálogo completo de ingredientes.
        **Roles permitidos**: `admin`, `chef`, `inventory_manager`
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de ingredientes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ingredient'

    post:
      tags:
        - Ingredientes
      summary: Crear nuevo ingrediente
      description: |
        Crea un nuevo ingrediente en el sistema.
        **Roles permitidos**: `admin`, `chef`
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngredientCreate'
      responses:
        '201':
          description: Ingrediente creado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseMessage'
        '400':
          description: Parámetros inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /ingredients/{id}:
    get:
      tags:
        - Ingredientes
      summary: Obtener ingrediente por ID
      description: |
        Recupera un ingrediente según su identificador.
        **Roles permitidos**: `admin`, `chef`, `inventory_manager`
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Detalle del ingrediente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ingredient'
        '404':
          description: Ingrediente no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Ingredientes
      summary: Actualizar ingrediente
      description: |
        Modifica los datos de un ingrediente existente.
        **Roles permitidos**: `admin`, `chef`
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngredientCreate'
      responses:
        '200':
          description: Ingrediente actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseMessage'
        '404':
          description: Ingrediente no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Ingredientes
      summary: Eliminar ingrediente
      description: |
        Elimina un ingrediente del sistema.
        **Roles permitidos**: `admin`
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Ingrediente eliminado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseMessage'
        '404':
          description: Ingrediente no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /ingredients/{id}/price-history:
    get:
      tags:
        - Ingredientes
      summary: Historial de precios de un ingrediente
      description: |
        Recupera el historial de cambios de precio para un ingrediente.
        **Roles permitidos**: `admin`, `supplier_manager`
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lista de cambios de precio
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PriceHistoryItem'
  /ingredients/{id}/allergens:
    get:
      tags:
        - Ingredientes
      summary: Obtener alérgenos de un ingrediente
      description: |
        Recupera todos los alérgenos asociados a un ingrediente específico.
        **Roles permitidos**: `admin`, `chef`, `waiter`
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lista de alérgenos del ingrediente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Allergen'
        '404':
          description: Ingrediente no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Ingredientes
      summary: Asignar alérgenos a ingrediente
      description: |
        Asigna múltiples alérgenos a un ingrediente.
        **Roles permitidos**: `admin`, `chef`
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - allergen_ids
              properties:
                allergen_ids:
                  type: array
                  items:
                    type: integer
      responses:
        '201':
          description: Alérgenos asignados correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseMessage'
        '400':
          description: Parámetros inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /ingredients/{id}/allergens/{allergen_id}:
    delete:
      tags:
        - Ingredientes
      summary: Quitar alérgeno de ingrediente
      description: |
        Elimina la asociación de un alérgeno específico con un ingrediente.
        **Roles permitidos**: `admin`, `chef`
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: path
          name: allergen_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Alérgeno eliminado del ingrediente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseMessage'
        '404':
          description: Ingrediente o alérgeno no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /recipes/{id}/ingredients:
    get:
      tags:
        - Recetas
      summary: Listar ingredientes de receta
      description: |
        Obtiene los ingredientes de una receta.
        **Roles permitidos**: `admin`, `chef`
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Ingredientes de la receta
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecipeIngredient'
        '404':
          description: No se encontraron ingredientes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /suppliers/{id}/ingredients:
    get:
      tags:
        - Proveedores
      summary: Listar ingredientes de proveedor
      description: |
        Obtiene los ingredientes asociados a un proveedor.
        **Roles permitidos**: `admin`, `supplier_manager`
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Ingredientes del proveedor
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SupplierIngredientItem'

    post:
      tags:
        - Proveedores
      summary: Asignar ingrediente a proveedor
      description: |
        Asocia un ingrediente a un proveedor.
        **Roles permitidos**: `admin`, `supplier_manager`
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupplierIngredientCreate'
      responses:
        '201':
          description: Ingrediente asignado al proveedor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseMessage'
        '400':
          description: Ingrediente ya asociado o parámetros inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /suppliers/{id}/ingredients/{ingredient_id}:
    put:
      tags:
        - Proveedores
      summary: Actualizar relación proveedor-ingrediente
      description: |
        Actualiza los datos de la relación entre un proveedor y un ingrediente (precio, tiempo de entrega, proveedor preferido).
        **Roles permitidos**: `admin`, `supplier_manager`
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: path
          name: ingredient_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupplierIngredientCreate'
      responses:
        '200':
          description: Relación actualizada correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseMessage'
        '404':
          description: Proveedor o ingrediente no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Proveedores
      summary: Desvincular ingrediente de proveedor
      description: |
        Elimina la asociación de un ingrediente con un proveedor.
        **Roles permitidos**: `admin`, `supplier_manager`
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: path
          name: ingredient_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Ingrediente eliminado del proveedor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseMessage'
        '404':
          description: Ingrediente o proveedor no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...